+++ rock/ts-sdk/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.1] - 2026-03-28
+
+### Fixed
+
+- **OSS HTTPS Connection** - OSS client now uses HTTPS by default
+  - Added `secure: true` to `ali-oss` client initialization in `setupOss()`
+  - Previously, OSS client defaulted to HTTP protocol, causing connection refused errors
+  - OSS buckets typically require HTTPS connections for security
+
 ## [1.3.0] - 2026-03-28
 
 ### Added